Fix python-future import
Move AddNearestThirdPartyFoldersToSysPath to server_utils.
Use PathToAllParentFolders instead of AncestorFolders in Javascript
completer. Manage the case where the path is a folder. Add a test for
this case.
Hardcode the third_party path in the Rust completer instead of using
PathToNearestThirdPartyFolder function.
