+++ ycmd/__main__.py
@@ -19,9 +19,7 @@ from __future__ import absolute_import
 from __future__ import unicode_literals
 from __future__ import print_function
 from __future__ import division
-from future import standard_library
-standard_library.install_aliases()
-from builtins import *  # noqa
+# Other imports from `future` must be placed after SetUpPythonPath.
 
 import sys
 import os
@@ -29,6 +27,10 @@ import os
 sys.path.insert( 0, os.path.dirname( os.path.abspath( __file__ ) ) )
 from server_utils import SetUpPythonPath, CompatibleWithCurrentCoreVersion
 SetUpPythonPath()
+
+from future import standard_library
+standard_library.install_aliases()
+from builtins import *  # noqa
 
 import sys
 import logging
