+++ ycmd/server_utils.py
@@ -31,20 +31,17 @@ CORE_NOT_COMPATIBLE_MESSAGE = (
   'ycmd can\'t run: ycm_core lib too old, PLEASE RECOMPILE'
 )
 
-
-def DirectoryOfThisScript():
-  return os.path.dirname( os.path.abspath( __file__ ) )
+DIR_OF_CURRENT_SCRIPT = os.path.dirname( os.path.abspath( __file__ ) )
 
 
 def SetUpPythonPath():
-  sys.path.insert( 0, os.path.join( DirectoryOfThisScript(), '..' ) )
+  sys.path.insert( 0, os.path.join( DIR_OF_CURRENT_SCRIPT, '..' ) )
 
-  from ycmd import utils
-  utils.AddNearestThirdPartyFoldersToSysPath( __file__ )
+  AddNearestThirdPartyFoldersToSysPath( __file__ )
 
 
 def ExpectedCoreVersion():
-  filepath = os.path.join( DirectoryOfThisScript(), '..', VERSION_FILENAME )
+  filepath = os.path.join( DIR_OF_CURRENT_SCRIPT, '..', VERSION_FILENAME )
   with io.open( filepath, encoding = 'utf8' ) as f:
     return int( f.read() )
 
@@ -56,3 +53,45 @@ def CompatibleWithCurrentCoreVersion():
   except AttributeError:
     return False
   return ExpectedCoreVersion() == current_core_version
+
+
+def AncestorFolders( path ):
+  folder = os.path.normpath( path )
+  while True:
+    parent = os.path.dirname( folder )
+    if parent == folder:
+      break
+    folder = parent
+    yield folder
+
+
+def PathToNearestThirdPartyFolder( path ):
+  for folder in AncestorFolders( path ):
+    path_to_third_party = os.path.join( folder, 'third_party' )
+    if os.path.isdir( path_to_third_party ):
+      return path_to_third_party
+  return None
+
+
+def AddNearestThirdPartyFoldersToSysPath( filepath ):
+  path_to_third_party = PathToNearestThirdPartyFolder( filepath )
+  if not path_to_third_party:
+    raise RuntimeError(
+        'No third_party folder found for: {0}'.format( filepath ) )
+
+  # NOTE: Any hacks for loading modules that can't be imported without custom
+  # logic need to be reproduced in run_tests.py as well.
+  for folder in os.listdir( path_to_third_party ):
+    # python-future needs special handling. Not only does it store the modules
+    # under its 'src' folder, but SOME of its modules are only meant to be
+    # accessible under py2, not py3. This is because these modules (like
+    # `queue`) are implementations of modules present in the py3 standard
+    # library. So to work around issues, we place the python-future last on
+    # sys.path so that they can be overriden by the standard library.
+    if folder == 'python-future':
+      folder = os.path.join( folder, 'src' )
+      sys.path.append( os.path.realpath( os.path.join( path_to_third_party,
+                                                       folder ) ) )
+      continue
+    sys.path.insert( 0, os.path.realpath( os.path.join( path_to_third_party,
+                                                        folder ) ) )
