+++ ycmd/utils.py
@@ -221,46 +221,16 @@ def TerminateProcess( pid ):
     os.kill( pid, signal.SIGTERM )
 
 
-def AncestorFolders( path ):
-  folder = os.path.abspath( path )
+def PathsToAllParentFolders( path ):
+  folder = os.path.normpath( path )
+  if os.path.isdir( folder ):
+    yield folder
   while True:
     parent = os.path.dirname( folder )
     if parent == folder:
       break
     folder = parent
     yield folder
-
-
-def PathToNearestThirdPartyFolder( path ):
-  for folder in AncestorFolders( path ):
-    path_to_third_party = os.path.join( folder, 'third_party' )
-    if os.path.isdir( path_to_third_party ):
-      return path_to_third_party
-  return None
-
-
-def AddNearestThirdPartyFoldersToSysPath( filepath ):
-  path_to_third_party = PathToNearestThirdPartyFolder( filepath )
-  if not path_to_third_party:
-    raise RuntimeError(
-        'No third_party folder found for: {0}'.format( filepath ) )
-
-  # NOTE: Any hacks for loading modules that can't be imported without custom
-  # logic need to be reproduced in run_tests.py as well.
-  for folder in os.listdir( path_to_third_party ):
-    # python-future needs special handling. Not only does it store the modules
-    # under its 'src' folder, but SOME of its modules are only meant to be
-    # accessible under py2, not py3. This is because these modules (like
-    # `queue`) are implementations of modules present in the py3 standard
-    # library. So to work around issues, we place the python-future last on
-    # sys.path so that they can be overriden by the standard library.
-    if folder == 'python-future':
-      folder = os.path.join( folder, 'src' )
-      sys.path.append( os.path.realpath( os.path.join( path_to_third_party,
-                                                       folder ) ) )
-      continue
-    sys.path.insert( 0, os.path.realpath( os.path.join( path_to_third_party,
-                                                        folder ) ) )
 
 
 def ForceSemanticCompletion( request_data ):
